requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.1
Flask==3.0.0
